refactor(core): extract apply_fusion_and_reranking to rag_retrieval (#459)
Closes #459. First slice (PR-H1a) of the rag_core.py post-retrieval
decomposition.

Moves four functions from rag_core.py (1933-2144) to new
rag_retrieval.py:
- apply_fusion_and_reranking — RRF (hybrid 2-way / m3 3-way) + cross-
  encoder rerank dispatch (rag_reranker.default_reranker call kept
  lazy)
- _coverage_counts — private helper for comparison balance
- apply_comparison_balance — coverage-aware top-k that guarantees
  min_per_target per comparison target before filling by global score
- reassemble_parent_sections — hierarchical retrieval mode

Circular-import avoidance: comparison_targets_for_analysis,
normalize_regions, normalize_page_span stay in rag_core (called from
many non-retrieval sites). rag_retrieval lazy-imports them inside the
functions that need them so the dependency graph remains
rag_retrieval -> rag_core (one-way).

rag_core re-exports the three public names so any existing caller
that imported them from rag_core keeps working without change.

Updates:
- CLAUDE.md repository map entry for rag_retrieval.py + the rag_reranker
  reference now points to rag_retrieval.apply_fusion_and_reranking
- ADR 0026 cross-ref points to rag_retrieval module
- scripts/_governance.py LOAD_BEARING_PATHS adds rag_retrieval.py
- .github/pull_request_template.md "load-bearing" hint updated to
  include rag_retrieval.py (so tests/test_governance.py
  test_pr_template_mentions_all_canonical_entries stays green)

Out of scope (PR-H1b follow-up):
- retrieve_candidates (1743-1932) + BM25/RRF helpers (2179-2313).
  Larger extraction touching ADR 0023 cross-refs (HyDE references
  retrieve_candidates line directly).

Verification: full local pytest 834 passed (1 skipped, 0 failed). The
two regression gates (test_naive_baseline_ranking_invariance,
test_retrieval_loop_regression) pass — JSON-identity preserved.
